Remove expensive index on Link. Simplify closure table again - 170 solid
Clean up prefer query, removing the depth from the temp table.  Add
index on unqueried.  Remove uniqueness constraint in closure as well.

The removal of the LINK.HASH index, while good for throughput, makes
things currently suck for producing DOT files for analyzing the DAG
graph for large graphs it seems.  Will see what I can do to fix it.

The simplification of CLOSURE is to simply use a BOOLEAN for
representing whether the row is the loop back to the parent (FALSE) or
the actual closure of the PARENT (TRUE).  No need to keep around depth
information, which is usually inaccurate due to the merge optimization
used by Apollo.