Transform to only use Avalanche gossip functional, but not working
Interim commit

Still have problems voting yes for txns for some reason.  There's a lot
of unknown txns in the parents, so the mixing doesn't seem to be
happening in skew with the parent selections.  Probably because I'm
parenting confidence 0 nodes more often than not, and so haven't been
flooded across the members by the time we query.  Or something else ;)