Moar excellent work on byzantine causal broadcast.
Have now integrated the basic bcb with the bounded buffer.  Optimize
reconcilliation using youngest age messages first. Message age is merged
when receiving messages and is incremented every gossip round of the
messenger.  The max age of a message is the Time To Live of the
Context's fireflies rings (diameter * 2 * k + 1).