Nice first real cut of CHOAM.
View changes are now performed inline, using the transactions of the
view to come to consensus on the next view ;)

Fixed a lot of issues, streamlined a lot of locks and executions - more
remaining, no doubt.

No rate limiting, etc, on txn submission - that's next on the plate.

Cleand up state machines in CHOAM and simplified the leaves.  Can likely
do more.

View reconfiguration ala Genesis is broken out and is planned for future
recovery regeneration of the view in "we're all dead, Jim" scenarios n'
such.

Still need to implement the equivalent of SYNCH from Consortium, to
force rebroadcast of each member's state for liveness.  That will be
tricky.