Fixed ch-rbc protocol errors, cleaned up tests, etc
created units subject to same ch-rbc as units from other processes.

Still have 2 flappers: a) preblocks differ in data at some mid point b)
stall in closing.

b appears to be due (in the use of the rbc module) to just random crap,
and the chain produced by the stalled unit matches, but it has not
transitioned the epoch.  Annoying.

a is undoubtedly related, and is likely due to some logic bomb I haven't
defused as of yet.