Refactor to sane multi-signature support.
Geebus.  Good stuff, though.

Big plug because Signer and Verifier sit at the bottom and are used
everywhere, for obvious reasons.

SigningThreshold is enforced, and signing/verifying in general should
now be multisig transparent.

This simplifies chain validation and this significantly clarified and
simplified the event receipt and witness structure for the KERL