Refactoring of metrics (#73)
* fix error handling logic around status - oy.

Back off rbc and producer gossip to 50 ms in ff testing in model.  Still
have to figure out why high frequency polling does not deal with the
disco. Need to winnow thread usage in all testing - I suspect this is
causing issues and needs to be cleaned up anyway

* single thread tests for squashing the rare flappers.

* scoping

* Looking good for smoke testing, but...

Need to really focus on txn submission logic :(

But, I knew that already ;)

Suprised that it works quite well without such ;)

* massive refactoring of metrics

* add warning of expected exceptions in testing

* fix ring test

* Prefix metrics with context digest

* firefly metrics prefixing with context id