squash that bug.
ultimately, the timing round was not decided correctly. The previous
method just took whatever units were on the level, and used the prefix
size, which is 2. For quorums of 4, this is still less than 2f+1 (3).

Worse, sometimes these samples - because of deterministic mixing of
order on rounds - would be size 1 instead of size 2.

The fix is to always use the first 2f+1 of the permutation, and require
2f+1 units in that permutation.

This was "caused" by the logic not checking for 2f+1 exact.

This was a random failure because most of the time it still works, when
in close synchronization and so would only show up when the nodes' views
are out of synch, due to natural variation.  Which is hard to produce in
the current, single process testing, as things are very uniform for the
most part.  lol.