chRBC gossip for Ethereal (#89)
* first sketch of RBC gossip

* Interim. Model after Adder.

Pull out nodes, generisize via Node interface.

* okay

* Better refactoring

* basic skeleton

* unit test dealing round (round 0) for RbcAdder

* test dealing for all members, full round 1

* fix produce logic, test to 3 rounds

* The traditional pre Easter code reformating

* doggone it  Missed this

* units produced by the node don't wait

* fix test logic for adder. test gossip + adder for 4 members

Woo hoo!

* Renaming and cleanup

* Refactoring phase 1

Move replicated state to the adder, in preparation for epoch
refactoring.

Remove adder test, as this is now reasonably complicated 3 phase gossip,
that is quite tedious to test - sue me; it was scafolding to get to this
point.

* refactor to include epoch

* Interim

Lotsa work

* interim

* squash that bug.

ultimately, the timing round was not decided correctly. The previous
method just took whatever units were on the level, and used the prefix
size, which is 2. For quorums of 4, this is still less than 2f+1 (3).

Worse, sometimes these samples - because of deterministic mixing of
order on rounds - would be size 1 instead of size 2.

The fix is to always use the first 2f+1 of the permutation, and require
2f+1 units in that permutation.

This was "caused" by the logic not checking for 2f+1 exact.

This was a random failure because most of the time it still works, when
in close synchronization and so would only show up when the nodes' views
are out of synch, due to natural variation.  Which is hard to produce in
the current, single process testing, as things are very uniform for the
most part.  lol.

* removed unused config, general sweeping up.

increase chrbc test to 13 members.

* use fixed number of random bloom filters per adder

Don't recreate the BFs each and every phase.

only care about updating >= current epoch

* interim

* finally figured out the problem.

Lots of clean up, adding logging, adding testing.

Problem was I was passing the parent, and not the child when the parent
is output and the child now has no waiting or missing parents.  lol

* Behold the new regime

* update README