Finally.  Fix permutation in Extender. Let's see if this fixes ye things
Geebus. The permutation mechanism I was using was causing the split
decisions. The previous mechanism broke down when there wasn't the same
number of units on the level to be decided.

I am an idiot.

Fixed a bunch of other stuff.  Tightened up quite a bit.  Still have an
issue with the last epoch not finishing due to timing round stuff.  Will
work on that.

Disable Model functional testing until I get txn retry logic in the
sharded oracle ;)