Bootstrapping Fireflies.
Revert back to olden daze where the EstablishmentEvent was used in ye Note. Use this as the verifier for the Participant who's Note this is.

This locks the View into the current Note of the Participant, as there's only that event as the Verifier.

Validation is used in the Gateway admission, to ensure only valid member IDs join.